Treat public DLs the same way as private ones in auto completion.
At least in our Exchange setup, the email addresses present in the
contact element for distribution lists aren't meaningful; the only valid
one is contained in the mailbox element. As the handling in
exchangeAbDistListDirectory.js already treats PublicDL and PrivateDL in
the same way, handle it like that in exchangeAbFolderDirectory as well.